Enhanced password management in Home.razor
Implemented Dependency Injection for the IJSRuntime interface in the Home.razor component to enable JavaScript interop, specifically for functionalities like copying text to the clipboard. Enhanced the copy to clipboard feature by adding an @onclick event to the password copy button, which now triggers the CopyPasswordToClipboard method. Modified the saved passwords list to display passwords in their added order and included interactive buttons for copying and deleting passwords. Introduced a deletion functionality for saved passwords with the HandleDeleteButtonClick method. Updated the CopyPasswordToClipboard method to accept a nullable string, allowing for more flexible copying actions. Added success notifications for copying and deleting actions using the ToastService, improving user feedback and experience.